[#42] common: Replace IRNode type and InnerRingInvoker func to ir.go file
Remove empty invoke.go source file.
